# Conflicts:
#	lib/eirinchan/live_vichan_import.ex
#	lib/eirinchan/runtime/config.ex
#	lib/eirinchan_web/controllers/board_html/catalog.html.heex
#	lib/eirinchan_web/controllers/page_controller.ex
#	lib/eirinchan_web/controllers/page_html/banners.html.heex
#	lib/eirinchan_web/controllers/page_html/flag.html.heex
#	lib/eirinchan_web/controllers/page_html/home.html.heex
#	lib/eirinchan_web/controllers/page_html/not_found.html.heex
#	lib/eirinchan_web/error_pages.ex
#	lib/eirinchan_web/plugs/fetch_theme.ex
#	lib/eirinchan_web/public_controller_helpers.ex
#	lib/mix/tasks/eirinchan.import_live_page.ex
#	test/eirinchan_web/controllers/page_controller_test.exs

+++ lib/eirinchan/boards.ex
@@ -7,6 +7,7 @@ defmodule Eirinchan.Boards do
 
   alias Eirinchan.Boards.{Board, EctoStore}
   alias Eirinchan.Boards.BoardRecord
+  alias Eirinchan.PrimaryBoard
   alias Eirinchan.Repo
   alias Eirinchan.Runtime
   alias Eirinchan.Runtime.{Config, RequestContext}
@@ -58,10 +59,18 @@ defmodule Eirinchan.Boards do
   end
 
   @spec delete_board(BoardRecord.t(), keyword()) ::
-          {:ok, BoardRecord.t()} | {:error, Ecto.Changeset.t()}
+          {:ok, BoardRecord.t()} | {:error, Ecto.Changeset.t() | :primary_board}
   def delete_board(%BoardRecord{} = board, opts \\ []) do
     repo = Keyword.get(opts, :repo, Repo)
-    repo.delete(board)
+
+    instance_config =
+      Keyword.get_lazy(opts, :instance_config, &Settings.effective_instance_config/0)
+
+    if board.uri == PrimaryBoard.uri(instance_config) do
+      {:error, :primary_board}
+    else
+      repo.delete(board)
+    end
   end
 
   @spec change_board(BoardRecord.t(), map()) :: Ecto.Changeset.t()
